backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
beautifulsoup4
python-dotenv
groq